apps/server/pyproject.toml
[project]
name = "doc-translator"
version = "0.1.0"
description = "A translator ai project for documentation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.117.1",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "marked>=0.9.1",
    "ollama>=0.6.0",
    "openai>=1.109.1",
    "pydantic-settings>=2.11.0",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
]

[tool.uv.pip]
upgrade = true

[dependency-groups]
dev = [
    "ruff>=0.13.2",
    "tomli-w>=1.2.0",
]
